requirements.txt
Django==1.11.6
django-cors-headers==2.1.0
django-environ==0.4.4
django-filter==1.0.4
djangorestframework==3.7.0
model-mommy==1.4.0
psycopg2==2.7.3.1
pytz==2017.2
six==1.11.0
